fix: restore chatbot handler
